Fix ANN delete on quantized Torch stores and tiny GGML indexes
- Torch delete assigned float32 zeros, which fails on uint8 (quantize: 1) and float16 stores. Zeros now match the array data type.
- GGML count, delete and append read shape[1], which a single row tensor doesn't have. A rows() helper reads the row count directly.
- GGML save then load of an empty index crashed since GGUF can't store a zero-row tensor. Save keeps the dimensions and load rebuilds the empty tensors, so it round-trips to an empty index like the other dense backends.

+++ src/python/txtai/ann/dense/numpy.py
@@ -63,8 +63,8 @@ class NumPy(ANN):
         # Filter any index greater than size of array
         ids = [x for x in ids if x < self.backend.shape[0]]
 
-        # Clear specified ids
-        self.backend[ids] = self.tensor(self.zeros((len(ids), self.backend.shape[1])))
+        # Clear specified ids, zeros must match the array data type (e.g. uint8 for quantized data)
+        self.backend[ids] = self.tensor(self.zeros((len(ids), self.backend.shape[1]), dtype=self.backend.dtype))
 
     def search(self, queries, limit):
         if self.qbits:
